Move errors/exceptions to errors module
