+++ django_inbound_email/backends/__init__.py
@@ -4,21 +4,6 @@ from importlib import import_module
 from django.conf import settings
 
 
-class RequestParseError(Exception):
-    """Error raised when the inbound request could not be parsed."""
-    pass
-
-
-class AttachmentTooLargeError(Exception):
-    """Error raised when an attachment is too large."""
-
-    def __init__(self, email, filename, size):
-        super(AttachmentTooLargeError, self)
-        self.email = email
-        self.filename = filename
-        self.size = size
-
-
 def get_backend_class():
     """Return reference to the configured backed class."""
     # this will (intentionally) blow up if the setting does not exist
