+++ django_inbound_email/errors.py
@@ -1,0 +1,16 @@
+
+class RequestParseError(Exception):
+    """Error raised when the inbound request could not be parsed."""
+    pass
+
+
+class AttachmentTooLargeError(Exception):
+    """Error raised when an attachment is too large."""
+
+    def __init__(self, email, filename, size):
+        super(AttachmentTooLargeError, self)
+        self.email = email
+        self.filename = filename
+        self.size = size
+
+
